requirements.txt
# Core ADK + Gemini
google-adk>=0.1.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.16.0

# Backend
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0

# Database
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# Knowledge base / embeddings
numpy==1.26.4
scikit-learn>=1.3.0

# Utilities
python-multipart>=0.0.5
aiohttp>=3.12.0
requests>=2.31.0
pytz>=2023.3
